// packages/model/src/unit/system/strategy/types/SystemHandlersTypes.ts
export type SystemMessageValue = string | number | SystemMessage[];

export type SystemMessage = {
  value?: SystemMessageValue;
  header?: string;
  sort?: string;
  component?: string;
  props?: unknown;
};